% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputDB.R
\name{inputDB}
\alias{inputDB}
\title{inputDB함수 사용법}
\usage{
inputDB(DB_VA, DB_DA)
}
\arguments{
\item{DB_VA, }{DB_DA}
}
\value{
표면적인 결과물은 없지만, 이 함수를 실행하지 않으면 분석이 시행되지 않습니다
}
\description{
inputDB함수는 계수DB를 입력하는 함수입니다
}
\examples{
inputpopperc("pop_perc.csv")
}
