% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputpopperc.R
\name{inputpopperc}
\alias{inputpopperc}
\title{inputpopperc함수 사용법}
\usage{
inputpopperc(DB_pop_perc)
}
\arguments{
\item{DB_pop_perc}{사전에 준비된 pop_perc 엑셀파일명 입력변수입니다}
}
\value{
표면적인 결과물은 없지만, 이 함수를 실행하지 않으면 분석이 시행되지 않습니다
}
\description{
inputpopperc함수는 타겟별 Impression 배분비율을 적용하는 함수입니다
}
\examples{
inputpopperc("pop_perc.csv")
}
