% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_reach_vehicle.R
\name{mix_reach_vehicle}
\alias{mix_reach_vehicle}
\title{mix_Reach_vehicle}
\usage{
mix_reach_vehicle(mix)
}
\arguments{
\item{mix, }{mix which result of mix_Reach}
}
\value{
data.frame
}
\description{
Copyright : DMC Media Data RnD team
version : 1.190719
}
\examples{
mix_reach(mix)
}
